refactor: RestaurantReviewQueryResponse -> RestaurantReviewsQueryResponse 네이밍 변경과, 내부 클래스로 dto 작성 (#583)
