+++ backend/src/main/java/com/celuveat/restaurant/query/RestaurantReviewQueryService.java
@@ -1,7 +1,7 @@
 package com.celuveat.restaurant.query;
 
-import com.celuveat.restaurant.query.dao.RestaurantReviewQueryResponseDao;
-import com.celuveat.restaurant.query.dto.RestaurantReviewQueryResponse;
+import com.celuveat.restaurant.query.dao.RestaurantReviewsQueryResponseDao;
+import com.celuveat.restaurant.query.dto.RestaurantReviewsQueryResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,9 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class RestaurantReviewQueryService {
 
-    private final RestaurantReviewQueryResponseDao restaurantReviewQueryResponseDao;
+    private final RestaurantReviewsQueryResponseDao restaurantReviewsQueryResponseDao;
 
-    public RestaurantReviewQueryResponse findAllByRestaurantId(Long restaurantId, Long memberId) {
-        return restaurantReviewQueryResponseDao.findAllByRestaurantId(restaurantId, memberId);
+    public RestaurantReviewsQueryResponse findAllByRestaurantId(Long restaurantId, Long memberId) {
+        return restaurantReviewsQueryResponseDao.findByRestaurantId(restaurantId, memberId);
     }
 }
