+++ backend/src/test/java/com/celuveat/acceptance/restaurant/RestaurantReviewAcceptanceSteps.java
@@ -6,8 +6,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.celuveat.restaurant.presentation.dto.SaveReviewRequest;
 import com.celuveat.restaurant.presentation.dto.UpdateReviewRequest;
-import com.celuveat.restaurant.query.dto.RestaurantReviewQueryResponse;
-import com.celuveat.restaurant.query.dto.RestaurantReviewSingleQueryResponse;
+import com.celuveat.restaurant.query.dto.RestaurantReviewsQueryResponse;
+import com.celuveat.restaurant.query.dto.RestaurantReviewsQueryResponse.RestaurantReviewSingleQueryResponse;
 import io.restassured.builder.MultiPartSpecBuilder;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -33,8 +33,8 @@ public class RestaurantReviewAcceptanceSteps {
     }
 
     // TODO : 이건 무슨...
-    public static RestaurantReviewQueryResponse 예상_응답(Long 말랑_아이디, Long 로이스_아이디, Long 도기_아이디) {
-        return new RestaurantReviewQueryResponse(4, List.of(
+    public static RestaurantReviewsQueryResponse 예상_응답(Long 말랑_아이디, Long 로이스_아이디, Long 도기_아이디) {
+        return new RestaurantReviewsQueryResponse(4, List.of(
                 new RestaurantReviewSingleQueryResponse(null, 말랑_아이디, "말랑", "https://말랑.jpg", "리뷰4", null, 0, false, 5.0, null),
                 new RestaurantReviewSingleQueryResponse(null, 말랑_아이디, "말랑", "https://말랑.jpg", "리뷰3", null, 0, false, 5.0, null),
                 new RestaurantReviewSingleQueryResponse(null, 로이스_아이디, "로이스", "https://로이스.jpg", "리뷰2", null, 0, false, 5.0, null),
@@ -129,9 +129,9 @@ public class RestaurantReviewAcceptanceSteps {
 
     public static void 응답을_검증한다(
             ExtractableResponse<Response> 응답,
-            RestaurantReviewQueryResponse 예상_응답
+            RestaurantReviewsQueryResponse 예상_응답
     ) {
-        RestaurantReviewQueryResponse responseBody = 응답.as(RestaurantReviewQueryResponse.class);
+        RestaurantReviewsQueryResponse responseBody = 응답.as(RestaurantReviewsQueryResponse.class);
         assertThat(responseBody).usingRecursiveComparison()
                 .ignoringExpectedNullFields()
                 .isEqualTo(예상_응답);
